refactor(fe): bookmark 로직 및 코드 개선
