+++ apps/web/src/features/bookmark/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useToggleBookmark';
