--- apps/web/src/features/bookmark/services/BookmarkService/hooks/index.ts
@@ -1,4 +1,0 @@
-export * from './useCreateBookmark';
-export * from './useFetchIsBookmarked';
-export * from './useRemoveBookmark';
-export * from './useToggleBookmark';
