--- apps/web/src/features/bookmark/services/BookmarkService/hooks/useFetchIsBookmarked.ts
@@ -1,14 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-
-import type { RequestType } from '@/shared/utils';
-
-import { queries } from '../queries';
-import type { service } from '../service';
-
-export const useFetchIsBookmarked = (
-  params: RequestType<typeof service.getIsBookmarked>,
-) => {
-  const { data } = useQuery(queries.getIsBookmarked(params));
-
-  return { data };
-};
