+++ apps/web/src/features/bookmark/services/BookmarkService/index.ts
@@ -1,9 +1,7 @@
-import * as hooks from './hooks';
 import * as queries from './queries';
 import { service } from './service';
 
 export const BookmarkService = {
   ...service,
   ...queries,
-  ...hooks,
 };
